requirements.txt
allure-pytest==2.6.1
allure-python-commons==2.6.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
enum34==1.1.6
idna==2.8
lxml==4.3.3
more-itertools==6.0.0
namedlist==1.7
pluggy==0.9.0
py==1.8.0
pygame==1.9.4
pytest==4.3.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
